wrapping up refactor
