+++ src/main/java/beam/utils/gtfs/Operator.java
@@ -1,5 +1,5 @@
 
-package beam.gtfs;
+package beam.utils.gtfs;
 
 import com.google.gson.annotations.SerializedName;
