+++ src/main/java/beam/utils/gtfs/merging/ScheduleMerger.java
@@ -19,7 +19,7 @@
  * *********************************************************************** *
  */
 
-package beam.gtfs.merging;
+package beam.utils.gtfs.merging;
 
 import org.apache.log4j.Logger;
 import org.matsim.api.core.v01.Id;
